refactor(frontend): flat file struct
